demo/spring-demo/src/main/java/nemo/test_container/MemberService.java
package nemo.test_container;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
@Service
public class MemberService {
	private final MemberRepository memberRepository;

	@Transactional
	public Member save(Member member){
		return memberRepository.save(member);
	}

	@Transactional
	public Member fineOne(Long id){
		return memberRepository.findById(id).orElseThrow();
	}
}
